Fix refactored directories
