+++ src/styled-components/primary_button.module.scss
@@ -1,4 +1,4 @@
-@import '../colors';
+@import '../styles/colors';
 
 .primary {
   background-color: $primary;
