+++ src/styles/index.css
@@ -1,7 +1,7 @@
 @font-face {
   font-family: "acremedium";
-  src: url("assets/fonts/jonathan_ball_-_acre-medium-webfont.woff2") format("woff2"),
-    url("assets/fonts/jonathan_ball_-_acre-medium-webfont.woff") format("woff");
+  src: url("../assets/fonts/jonathan_ball_-_acre-medium-webfont.woff2") format("woff2"),
+    url("../assets/fonts/jonathan_ball_-_acre-medium-webfont.woff") format("woff");
   font-weight: normal;
   font-style: normal;
 }
